Guidance: Feature Flags (#160)
* guidance: Add guidance for feature flags

Feature flags are a crucial component in SCORE to handle
configuration and variant management.

Introduces guidance to document how SCORE works with
feature flags.

* guidance: Link to feature flags guidance

In the feature request template the author must provide
a feature flag name. To understand what this is about,
we link to the guidance, so that the author can read up
the topic before deciding on a name.

* doc: update copyright to current year

The PR took long enough to make it to 2025

Co-authored-by: Alexander Lanin <Alexander.Lanin@etas.com>
Signed-off-by: Ulrich Huber <ulrich@huberulrich.de>

---------

Signed-off-by: Ulrich Huber <ulrich@huberulrich.de>
Co-authored-by: Alexander Lanin <Alexander.Lanin@etas.com>